refactor: Move storage related logic to repositories and move business logic to services
