+++ Readme.md
@@ -1,0 +1,14 @@
+# go-fiber-auth
+
+This repository is created to show a basic Go app with user account creation and authentication built with fiber & a postgres db.The aim of the project is to learn how to setup a fiber app with authentication and authorization as well as setting up a clean project structure.
+
+## Key dependencies I used
+
+|Dependency | Version |
+|-----------|---------|
+| Fiber     |         |
+| Gorm      |
+| JWT       |
+| Validation|   
+| Postgres  |      
+
