+++ dto/user/create_user_dto.go
@@ -1,0 +1,8 @@
+package dto
+
+type SignupUserDataDto struct {
+	Username  string `json:"username" validate:"required"`
+	Password  string `json:"password" validate:"required"`
+	FirstName string `json:"first_name" validate:"required"`
+	LastName  string `json:"last_name" validate:"required"`
+}
