+++ dto/user/user_response_dto.go
@@ -1,0 +1,7 @@
+package dto
+
+type UserResponseDto struct {
+	Username  string `json:"username"`
+	FirstName string `json:"first_name"`
+	LastName  string `json:"last_name"`
+}
