+++ middleware/auth.go
@@ -2,18 +2,19 @@ package middleware
 
 import (
 	"fib/config"
-
 	jwtware "github.com/gofiber/contrib/jwt"
 	"github.com/gofiber/fiber/v2"
 )
 
 func Protected() fiber.Handler {
-	return jwtware.New(jwtware.Config{
-		SigningKey: jwtware.SigningKey{
-			Key: []byte(config.Config("SECRET")),
+	return jwtware.New(
+		jwtware.Config{
+			SigningKey: jwtware.SigningKey{
+				Key: []byte(config.Config("SECRET")),
+			},
+			ErrorHandler: jwtError,
 		},
-		ErrorHandler: jwtError,
-	})
+	)
 }
 
 func jwtError(c *fiber.Ctx, err error) error {
